refactor: extract OAuth endpoints into server/oauth.ts
Move the /.well-known/oauth-authorization-server and /oauth/token
request handlers out of main() in server/index.ts into a dedicated
module. server/index.ts drops from ~410 lines to ~185. Also replaces
regex .exec() calls with .match() throughout the HTTP handler for
consistency.
